Disabling realistic thrust for (some) NPC ships
Not entirely accurate due to spaghetti mess that is controlling ships.
First if any RC has AutomaticBehavior != null it's NPC, that should catch drones.
Next iterates players and if they control the flying cockpit then it's not NPC but if they control a turret on that grid and are in a seat on the same grid it will assume it's not NPC (because there's no reliable way to tell if their cockpit actually flies the ship).
And finally checks if the first bigowner is in a NPC-only faction for any uncontrolled or custom script controleld ships.

Known issues:
- Controlling a turret while in a cockpit on a NPC-owned ship will make it have realistic thrusts no matter who is actually flying it.